requirements.txt
Flask==2.1.0
flasgger==0.9.5
mysql-connector-python==8.1.0
Flask-Cors==3.0.10
